+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [8.0.6] - 2023-03-24
 
+### Fixed
+- Improved error message when using self-signed certificates
+  [cyberark/conjur-cli-go#119](https://github.com/cyberark/conjur-cli-go/pull/119)
+
 ## [8.0.5] - 2023-03-24
 
 ### Changed